% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtable_type.R
\name{as_hdtable_type}
\alias{as_hdtable_type}
\title{hdtable_type Vectors}
\usage{
as_hdtable_type(x)
}
\arguments{
\item{x}{object to be coerced}
}
\value{
attempts to coerce its argument to hdtable_type type
}
\description{
coerces its argument to a hdtable_type. It is an abbreviated form of hdtable_type.
}
\examples{

some_chr_value <- "Cat"
class(some_chr_value)

some_frt_value <- as_hdtable_type(some_chr_value)
class(some_frt_value)

}
