% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{as_hdtableType}
\alias{as_hdtableType}
\title{hdtableType Vectors}
\usage{
as_hdtableType(x)
}
\arguments{
\item{x}{object to be coerced}
}
\value{
attempts to coerce its argument to hdtableType type
}
\description{
coerces its argument to a hdtableType. It is an abbreviated form of hdtableType.
}
\examples{

some_chr_value <- "Cat"
class(some_chr_value)

some_frt_value <- as_hdtableType(some_chr_value)
class(some_frt_value)

}
