% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{expand_hdtable_type_group}
\alias{expand_hdtable_type_group}
\title{hdtable_type Vectors}
\usage{
expand_hdtable_type_group(hdtable_type_group)
}
\arguments{
\item{hdtable_type_group}{a grouped hdtable_type object}
}
\value{
a string value showing all the hdtable_type values of an object
}
\description{
Reverses the effect of \code{\link[=get_hdtable_type_group]{get_hdtable_type_group()}} and split every single hdtable_type from an object.
}
\examples{

x <- c("Cat-Num-Cat")
fr <- hdtable_type(x)
grouped_fr <- get_hdtable_type_group(fr)
expand_frGroup(grouped_fr)

}
