% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{expand_hdtableTypeGroup}
\alias{expand_hdtableTypeGroup}
\title{hdtableType Vectors}
\usage{
expand_hdtableTypeGroup(hdtableTypeGroup)
}
\arguments{
\item{hdtableTypeGroup}{a grouped hdtableType object}
}
\value{
a string value showing all the hdtableType values of an object
}
\description{
Reverses the effect of \code{\link[=get_hdtableTypeGroup]{get_hdtableTypeGroup()}} and split every single hdtableType from an object.
}
\examples{

x <- c("Cat-Num-Cat")
fr <- hdtableType(x)
grouped_fr <- get_hdtableTypeGroup(fr)
expand_frGroup(grouped_fr)

}
