% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtable_type.R
\name{get_hdtable_type_group}
\alias{get_hdtable_type_group}
\title{hdtable_type Vectors}
\usage{
get_hdtable_type_group(hdtable_type_str)
}
\arguments{
\item{hdtable_type_str}{a string value showing a grouped hdtable_types view}
}
\value{
a grouped view of given hdtable_types values
}
\description{
a grouped way of reading hdtable_types values
}
\examples{

x <- c("Cat-Num-Cat")
fr <- hdtable_type(x)
get_hdtable_type_group(fr)

}
