% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{get_hdtableTypeGroup}
\alias{get_hdtableTypeGroup}
\title{hdtableType Vectors}
\usage{
get_hdtableTypeGroup(hdtableType_str)
}
\arguments{
\item{hdtableType_str}{a string value showing a grouped hdtableTypes view}
}
\value{
a grouped view of given hdtableTypes values
}
\description{
a grouped way of reading hdtableTypes values
}
\examples{

x <- c("Cat-Num-Cat")
fr <- hdtableType(x)
get_hdtableTypeGroup(fr)

}
