% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtable-package.R
\docType{package}
\name{hdtable-package}
\alias{hdtable}
\alias{hdtable-package}
\title{hdtable: Homodatum Table}
\description{
Vector definitions for abstract data types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/hdtable}
  \item Report bugs at \url{https://github.com/datasketch/hdtable/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
