% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtable.R
\name{hdtable}
\alias{hdtable}
\title{Create a hdtable data frame}
\usage{
hdtable(
  d,
  dic = NULL,
  hdtable_type = NULL,
  name = NULL,
  description = NULL,
  slug = NULL,
  meta = NULL,
  formats = NULL,
  ...
)
}
\arguments{
\item{dic}{a custom variable dictionary can be added. \code{\link[=create_dic]{create_dic()}} can help you with that.}

\item{hdtable_type}{The type of hdtable to create}

\item{name}{a custom name can be added}

\item{slug}{a custom slug can be added. If not, hdtable will try creating one.}

\item{meta}{Custom Metadata can be added}

\item{x}{A data frame}

\item{nam}{a custom description can be added}
}
\value{
A hdtable object
}
\description{
Create a hdtable object from a data frame. The main value of a hdtable is its metadata. When creating it, hdtable will add to the data frame the following information:
\itemize{
\item data: original data frame data. When it is created, the hdtable will convert original variable R types onto homodatum ones (Num, Cat, Pct, etc. -see \code{\link[=available_hdtypes]{available_hdtypes()}})
\item dic: A diccionary is created with three variable characteristics: id, label and hdtype
\item hdtable_type: Shows all variable types based in homodatum schema
\item group: A grouped view of hdtable_type
\item name: Name for the hdtable data frame, setted on \emph{name} argument
\item description: Description for the hdtable data frame, setted on \emph{description} argument
\item slug: a custom slug can be added
\item stats: Depending on the variable type given by homodatum, the hdtable will generate different kind of statistics: nrow, ncol, n_unique, n_na, pct_na, min, max
}
}
\examples{
hdtable(mtcars, hdtable_type = "Num", name = "MTCars")

}
