% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{hdtable_type}
\alias{hdtable_type}
\title{hdtable_type Vectors}
\usage{
hdtable_type(x = character())
}
\arguments{
\item{x}{object to be created or coerced}
}
\value{
returns a hdtable_type value
}
\description{
Creates or coerces objects of type "hdtable_type"
}
\examples{

hdtable_type("Cat")

}
