% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{hdtable_type_group}
\alias{hdtable_type_group}
\title{hdtable_type_group of an hdtable}
\usage{
hdtable_type_group(x)
}
\arguments{
\item{x}{hdtable}
}
\value{
returns a string
}
\description{
Gets the hdtable_type_group from an hdtable
}
\examples{

hdtable_type("Cat")

}
