% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{hdtable_type_hdtypes}
\alias{hdtable_type_hdtypes}
\title{hdtable_type Vectors}
\usage{
hdtable_type_hdtypes(x, chr = FALSE)
}
\arguments{
\item{x}{an available hdtable_type value}
}
\value{
an hdtype value
}
\description{
convert hdtable_types value(s) into hdtype
}
\examples{

x <- hdtable_type("Cat")
class(x)

x_hdt <- hdtable_type_hdtypes(x)
class(x_hdt)

}
