% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{hdtableType}
\alias{hdtableType}
\title{hdtableType Vectors}
\usage{
hdtableType(x = character())
}
\arguments{
\item{x}{object to be created or coerced}
}
\value{
returns a hdtableType value
}
\description{
Creates or coerces objects of type "hdtableType"
}
\examples{

hdtableType("Cat")

}
