% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{hdTableType_hdTypes}
\alias{hdTableType_hdTypes}
\title{hdTableType Vectors}
\usage{
hdTableType_hdTypes(x, chr = FALSE)
}
\arguments{
\item{x}{an available hdTableType value}
}
\value{
an hdType value
}
\description{
convert hdTableTypes value(s) into hdType
}
\examples{

x <- hdTableType("Cat")
class(x)

x_hdt <- hdTableType_hdTypes(x)
class(x_hdt)

}
