% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{hdtableType_str}
\alias{hdtableType_str}
\title{hdtableType Vectors}
\usage{
hdtableType_str(x)
}
\arguments{
\item{x}{an available hdtableType value, hdtableType dataframe or hd_tbl dataframe}
}
\value{
a character value
}
\description{
convert hdtableTypes value(s) into character
}
\examples{

x <- hdtableType("Cat")
class(x)

x_chr <- hdtableType_str(x)
class(x_chr)

}
