% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtable.R
\name{is_hdtable}
\alias{is_hdtable}
\title{hdtable data frame}
\usage{
is_hdtable(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdtable or not.
}
\description{
test for objects of type "hdtable"
}
\examples{
some_df <- hdtable(mtcars)
is_hdtable(some_df)

}
