% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{is_hdtable_type}
\alias{is_hdtable_type}
\title{hdtable_type Vectors}
\usage{
is_hdtable_type(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdtable_type or not.
}
\description{
Creates or test for objects of type "hdtable_type"
}
\examples{

value <- hdtable_type("Cat")
is_hdtable_type(value)

}
