% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{is_hdTableType}
\alias{is_hdTableType}
\title{hdTableType Vectors}
\usage{
is_hdTableType(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdTableType or not.
}
\description{
Creates or test for objects of type "hdTableType"
}
\examples{

value <- hdTableType("Cat")
is_hdTableType(value)

}
