% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.R
\name{update_dic}
\alias{update_dic}
\title{Update Dictionary with data}
\usage{
update_dic(dic, d, stats = TRUE)
}
\arguments{
\item{dic}{The original dictionary}

\item{d}{The data set for which the user is creating the dictionary for}
}
\value{
a data frame with five columns: id, label, hdtype, format and stats
}
\description{
\code{update_dic()} Creates a data.frame dictionary identifying column
id (with cleaned variable names), label and homodatum variable type
}
\examples{
d <- mtcars
new_dic <- create_dic(d)
}
