% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{as_hdTableType}
\alias{as_hdTableType}
\title{hdTableType Vectors}
\usage{
as_hdTableType(x)
}
\arguments{
\item{x}{object to be coerced}
}
\value{
attempts to coerce its argument to hdTableType type
}
\description{
coerces its argument to a hdTableType. It is an abbreviated form of hdTableType.
}
\examples{

some_chr_value <- "Cat"
class(some_chr_value)

some_frt_value <- as_hdTableType(some_chr_value)
class(some_frt_value)

}
