% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{expand_hdTableTypeGroup}
\alias{expand_hdTableTypeGroup}
\title{hdTableType Vectors}
\usage{
expand_hdTableTypeGroup(hdTableTypeGroup)
}
\arguments{
\item{hdTableTypeGroup}{a grouped hdTableType object}
}
\value{
a string value showing all the hdTableType values of an object
}
\description{
Reverses the effect of \code{\link[=get_hdTableTypeGroup]{get_hdTableTypeGroup()}} and split every single hdTableType from an object.
}
\examples{

x <- c("Cat-Num-Cat")
fr <- hdTableType(x)
grouped_fr <- get_hdTableTypeGroup(fr)
expand_frGroup(grouped_fr)

}
