% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{get_hdTableTypeGroup}
\alias{get_hdTableTypeGroup}
\title{hdTableType Vectors}
\usage{
get_hdTableTypeGroup(hdTableType_str)
}
\arguments{
\item{hdTableType_str}{a string value showing a grouped hdTableTypes view}
}
\value{
a grouped view of given hdTableTypes values
}
\description{
a grouped way of reading hdTableTypes values
}
\examples{

x <- c("Cat-Num-Cat")
fr <- hdTableType(x)
get_hdTableTypeGroup(fr)

}
