% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTable-package.R
\docType{package}
\name{hdTable-package}
\alias{hdTable}
\alias{hdTable-package}
\title{hdTable: Homodatum Table}
\description{
Vector definitions for abstract data types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/hdTable}
  \item Report bugs at \url{https://github.com/datasketch/hdTable/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
