% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTable.R
\name{hdTable}
\alias{hdTable}
\title{Create a hdTable data frame}
\usage{
hdTable(
  d,
  dic = NULL,
  hdTableType = NULL,
  name = NULL,
  description = NULL,
  slug = NULL,
  meta = NULL
)
}
\arguments{
\item{dic}{a custom variable dictionary can be added. \code{\link[=create_dic]{create_dic()}} can help you with that.}

\item{hdTableType}{The type of hdTable to create}

\item{name}{a custom name can be added}

\item{slug}{a custom slug can be added. If not, hdTable will try creating one.}

\item{meta}{Custom Metadata can be added}

\item{x}{A data frame}

\item{nam}{a custom description can be added}
}
\value{
A hdTable object
}
\description{
Create a hdTable object from a data frame. The main value of a hdTable is its metadata. When creating it, hdTable will add to the data frame the following information:
\itemize{
\item data: original data frame data. When it is created, the hdTable will convert original variable R types onto homodatum ones (Num, Cat, Pct, etc. -see \code{\link[=available_hdTypes]{available_hdTypes()}})
\item dic: A diccionary is created with three variable characteristics: id, label and hdType
\item hdTableType: Shows all variable types based in homodatum schema
\item group: A grouped view of hdTableType
\item name: Name for the hdTable data frame, setted on \emph{name} argument
\item description: Description for the hdTable data frame, setted on \emph{description} argument
\item slug: a custom slug can be added
\item stats: Depending on the variable type given by homodatum, the hdTable will generate different kind of statistics: nrow, ncol, n_unique, n_na, pct_na, min, max
}
}
\examples{
hdTable(mtcars, hdTableType = "Num", name = "MTCars")

}
