% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{hdTableType}
\alias{hdTableType}
\title{hdTableType Vectors}
\usage{
hdTableType(x = character())
}
\arguments{
\item{x}{object to be created or coerced}
}
\value{
returns a hdTableType value
}
\description{
Creates or coerces objects of type "hdTableType"
}
\examples{

hdTableType("Cat")

}
