% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{hdtableType_hdtypes}
\alias{hdtableType_hdtypes}
\title{hdtableType Vectors}
\usage{
hdtableType_hdtypes(x, chr = FALSE)
}
\arguments{
\item{x}{an available hdtableType value}
}
\value{
an hdtype value
}
\description{
convert hdtableTypes value(s) into hdtype
}
\examples{

x <- hdtableType("Cat")
class(x)

x_hdt <- hdtableType_hdtypes(x)
class(x_hdt)

}
