% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTableType.R
\name{hdTableType_str}
\alias{hdTableType_str}
\title{hdTableType Vectors}
\usage{
hdTableType_str(x)
}
\arguments{
\item{x}{an available hdTableType value, hdTableType dataframe or hd_tbl dataframe}
}
\value{
a character value
}
\description{
convert hdTableTypes value(s) into character
}
\examples{

x <- hdTableType("Cat")
class(x)

x_chr <- hdTableType_str(x)
class(x_chr)

}
