% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTable.R
\name{is_hdTable}
\alias{is_hdTable}
\title{hdTable data frame}
\usage{
is_hdTable(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdTable or not.
}
\description{
test for objects of type "hdTable"
}
\examples{
some_df <- hdTable(mtcars)
is_hdTable(some_df)

}
