% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtableType.R
\name{is_hdtableType}
\alias{is_hdtableType}
\title{hdtableType Vectors}
\usage{
is_hdtableType(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdtableType or not.
}
\description{
Creates or test for objects of type "hdtableType"
}
\examples{

value <- hdtableType("Cat")
is_hdtableType(value)

}
